Create chess-knight-way.py
You need to output "YES", if the specified knight's move is correct, if the record is correct (in the sense of correctness of the coordinates), but the move is impossible, then you need to output "NO". If the coordinates are not defined or are set incorrectly, then display the message "ERROR".